Struct rimd::AbsoluteEvent

pub struct AbsoluteEvent {
    // some fields omitted
}

An AbsoluteEvent is an event that has an absolute time This is useful for apps that want to store events internally with absolute times and then quickly build an SMF file for saving etc...

Methods

impl AbsoluteEvent

fn new_midi(time: u64, midi: MidiMessage) -> AbsoluteEvent

fn new_meta(time: u64, meta: MetaEvent) -> AbsoluteEvent

fn is_midi(&self) -> bool

Return true if the event inside this AbsoluteEvent is a midi event, false if it's a meta event

fn is_meta(&self) -> bool

Return true if the event inside this AbsoluteEvent is a meta event, false if it's a midi event

fn get_event(&self) -> &Event

fn get_time(&self) -> u64