Struct rimd::SMF

pub struct SMF {
    pub format: SMFFormat,
    pub tracks: Vec<Track>,
    pub division: i16,
}

A standard midi file

Fields

format

The format of the SMF

tracks

Vector holding each track in this SMF

division

The unit of time for delta timing. If the value is positive, then it represents the units per beat. For example, +96 would mean 96 ticks per beat. If the value is negative, delta times are in SMPTE compatible units.

Methods

impl SMF

fn from_file(path: &Path) -> Result<SMF, SMFError>

Read an SMF file at the given path

fn from_reader(reader: &mut Read) -> Result<SMF, SMFError>

Read an SMF from the given reader